Add first optimization task
Uses the ENABLE_FAST_MUL configuration flag. In this task, the goal is to optimize the performance. The 'optimized' version is simply the module with ENABLE_FAST_MUL=1.

Note that this HDL code could be re-used for another task. Because setting ENABLE_FAST_MUL=1 increases the area, we could make another optimization target where the ENABLE_FAST_MUL=0 version is the 'optimized' module.

Right now the optimized and non-optimized code both pass all the testbench cases. We need to get metrics though for the improved performance or area.