Add synth with subprocess and re-use task1
Task 1 is re-used because if the fast matmul is enabled, then the performance goes up, but so does the area. We re-use the exact same code, but from two different perspectives.

Because of this, I just created symbolic links to the task1 code. Not super clean, but I wanted to avoid duplicated code.