Add tasks 3-5 and debug output
All of the tasks concern either the multiply or the divide module in the pico project.

These are not the most exciting modules to optimize, but they are pretty simple to benchmark. Optimizing the entire RISC-V CPU is quite difficult because the architecture is pretty much just one giant module. This helps out when plopping it into a HDL project, but is more finnicky when trying to get a clear reference/optimized version for benchmarking.

Added some debug statements to the python processes.

Also note that I grouped all the "optimized" baseline codes into a single folder, instead of breaking it up into human/ and llm/ folders. I did this because the LLM only optimized some of the tasks, and the directory structure was a bit messy because of missing tasks for different optimization techniques.

Another small note -> I'm going to use mainly 3 baselines. The reference and the optimized code baselines, but then also Yosys synthesis optimization flags.

Finally, task #5 is to optimize the power consumption, but we currently have no tool which provides this information. That'll be the next commit.