Add subparsers and LLM code
Since the different steps of the module require different arguments, I used the subparser() method as explaind in the argparse docs.

I also added the initial LLM inference code. This needs to be wrapped in a loop.

I don't think we need the LLM to necessarily 'call' any tool. This is because we are only using the generated Verilog code as the parameter. We will just make sure to pass compilation, simulation, and synthesis before we give PPA feedback to the agent.

Also small fix on symlinks breaking from earlier commit