Fix loop due to os.chdir
Agentic loop now works as expected. The issue was that when I called the sim() in a loop we kept created a build directory and cd'ing into it.

I also added some delimters to indicate user/assistant/tool outputs. The agent code will eventually be refactored to be a class.